requirements.txt
# SafeScene AI - Python Requirements
# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
stable-baselines3>=2.0.0
transformers>=4.30.0

# GAN Implementation
lpips
pytorch-fid
kornia>=0.7.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
h5py>=3.9.0

# Web Interface
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.25.0
plotly>=5.15.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.0
